splitting into a library and a binary
