+++ src/lib.rs
@@ -4,20 +4,9 @@ use bitcoin::{Network, PublicKey};
 use std::convert::TryInto;
 use std::str::FromStr;
 
-struct ElectrumExtendedPubKey {
+pub struct ElectrumExtendedPubKey {
     xpub: ExtendedPubKey,
     kind: String,
-}
-
-fn main() -> Result<(), String> {
-    let mut args = std::env::args();
-    args.next(); // first is program name
-    let electrum_xpub = args
-        .next()
-        .ok_or_else(|| "You must specify an electrum xpub as first argument".to_string())?;
-    let electrum_xpub = ElectrumExtendedPubKey::from_str(&electrum_xpub)?;
-    println!("{:?}", electrum_xpub.to_descriptors());
-    Ok(())
 }
 
 impl FromStr for ElectrumExtendedPubKey {
